LayoutTests/http/tests/security/contentSecurityPolicy/1.1/securitypolicy-allowimagefrom.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
  <body>
    <script src="../resources/securitypolicy-tests-base.js"></script>
    <script>
      if (document.SecurityPolicy.allowsImageFrom('http://example.com/'))
          log('PASS image is allowed when no policy exists.');
      else
          log('FAIL image is not allowed when no policy exists.');

      injectPolicy("img-src http://notexample.com;");

      if (!document.SecurityPolicy.allowsImageFrom('http://example.com/'))
          log('PASS image is not allowed when policy exists.');
      else
          log('FAIL image is allowed when policy exists.');
    </script>
  </body>
</html>
