LayoutTests/http/tests/security/contentSecurityPolicy/1.1/securitypolicy-allowimagefrom-expected.txt

PASS image is allowed when no policy exists.
PASS image is not allowed when policy exists.
